pyproject.toml
[tool.poetry]
name = "qlient-aiohttp"
version = "0.1.0-beta.2"
description = "A blazingly fast and modern graphql client based on qlient-core and aiohttp"
authors = ["Daniel Seifert <info@danielseifert.ch>"]
maintainers = ["Daniel Seifert <info@danielseifert.ch>"]
readme = "README.md"

packages = [
    { include = "qlient", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
qlient-core = "*"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
coverage = "^6.4.1"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-aiohttp = "^1.0.4"
pytest-cov = "^3.0.0"
black = "^22.3.0"
strawberry-graphql = "^0.115.0"
mkdocs = "^1.3.0"
mkdocs-material = "^8.3.9"
mkdocstrings = { version = "^0.19.0", extras = ["python"] }
mkdocs-include-markdown-plugin = "^3.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
